+++ lib/src/doctor/doctor_runner.dart
@@ -88,6 +88,7 @@ class DoctorRunner {
       ],
       AdmobApiAccessCheck(),
       AdmobAppIdCheck(),
+      DartDefineFileCheck(),
     ];
   }